Audit fixes: real loyalty numbers everywhere, the QR the till scans, stock texts that reach people
The owner asked for audits of the loyalty programme, the inventory
reorder alerts and the customer accounts flow, then said "next". These
are the fixes in the order suggested; the points-expiry over-count waits
until expiry is switched on.

Loyalty. The customer row's points and tier were written once at
sign-up and never again, so the admin customer list, the POS search,
the till's scan result and the login payload showed zero and bronze for
everyone while the real balance lived on the loyalty account. Every
change to the account now copies the two values across, and a one-time
migration copies today's balances over. The nightly reconciliation
compared MySQL's string sum strictly to an integer and reported every
account as drift each night; it compares numbers now.

Customer accounts. The "My code" QR the till scans (owner, 2026-09-02)
sat inside the signed-out branch behind an "is signed in" check, so it
never rendered; it lives on the signed-in hub now. The account page's
loyalty panel was unreachable — no row opened it — and hardcoded tier
thresholds and a redeem rate the admin settings can change, so it is
gone; the Rewards page, which reads the live rates, is the one place.
The profile card shows spendable points, as Rewards does. A password
reset code asked for through the general OTP endpoint spent the login
budget, four times the forgot-password one for the same SMS; it spends
the reset budget now.

Inventory. The reorder text only mentioned alerts created that morning:
an item that hit its reorder point while snoozed was never texted when
the snooze ended, and one sitting below the point for weeks was never
mentioned again. Each open alert now remembers when it was last texted,
and the digest names every open alert not texted in a week — new,
snooze over, or still low. The expiry check printed to a console nobody
reads; it now texts the owners once a day, under the same "stock alert"
switch, about stock on hand that expires within the week.

Two migrations (a backfill and inventory_reorder_alerts.notified_at).
Rebuilt order and admin bundles.

+++ backend/app/Console/Commands/ReconcileLoyaltyBalances.php
@@ -25,10 +25,13 @@ class ReconcileLoyaltyBalances extends Command
         $mismatches = [];
 
         foreach ($accounts as $account) {
-            $ledgerSum = LoyaltyLedger::where('customer_id', $account->customer_id)
+            // MySQL hands SUM() back as a string; compared strictly to the
+            // integer balance every account read as drift and was
+            // "fixed" nightly to the value it already had (audit, 2026-09-24).
+            $ledgerSum = (int) LoyaltyLedger::where('customer_id', $account->customer_id)
                 ->sum('points');
 
-            if ($ledgerSum !== $account->points_balance) {
+            if ($ledgerSum !== (int) $account->points_balance) {
                 $mismatches[] = [
                     'customer_id' => $account->customer_id,
                     'balance' => $account->points_balance,
